refactor: extract metadata types and helpers into shared file
